[TASK] Simplify AssetProvider's PageRenderer usage
AssetProvider basically mirrored PageRenderer in allowing you to add
assets either inline, as file or as lib. With PageRenderer changing to
only render once, and more official ways to provide assets conditionally
than when this package was originally conceived, I'm downsizing its
capabalities with the intention of at some point eliminating the
package.

Currently, this means we're no longer going to mirror PageRenderer.
First step: Everything will end up as a file in either headerData or
footerData containers of the PageRenderer (same that are used by the
ActionController for template-driven assets). For now we're remaining
compatible with all original configurations however. A few configuration
options are sacrificed since we're not intending to copy over
PageRenderer capabilities that were inherent to inline/file/lib adding
only.